Reset `display_errors` later during execution.
The ideal time to do it is as the very first thing that happens when the file is loaded, in order to catch all errors. That works fine with the command is registered from an `mu-plugin`, but not if it's installed as a package.

When installed as a package, it gets executed before `apply_filters()` and `wp_debug_mode()` are available, and calling it will crash with a fatal error.